moved widgets to separate folder
