+++ widgets/highlighter/pyHighlight.py
